Add HTTP server to tests
The two existing tests currently requires an internet connection,
since the request is done towards a real HTTP server.

With this commit, the client will do requests towards the local HTTP
server.
The HTTP server is configured to respond with a choosen response, so that
it matches the expected response to the client.