Add FTP test
Added an FTP test. It starts a server, sends a request and receives
the response.

However, no verification is done. This will be added in a later stage
when the support is added in the FTP repository.

The inner classes will also be removed as soon as the FTP repository
adds methods in the Factory to create the instances.